Reorganize nlp tools
