+++ nlp/__init__.py
@@ -1,0 +1,2 @@
+from nlp.Parser import Parser
+from nlp.Tagger import Tagger
