--- parse/__init__.py
@@ -1,1 +1,0 @@
-from parse.Parser import Parser
